requirements.txt
flake8==6.0.0
pytest==7.2.1
psycopg2-binary==2.9.5
Jinja2==3.1.2
pep8-naming==0.13.3